UI Improvements and warning cleanup
- Allow the laps associated with a wrong person to be associated with the correct person in the EventViewer window of TrailMeisterViewer. This would help in situations where someone grabs the wrong tag or someone borrows a tag without updating the person associated with it before using it. The event happens with this tag assigned to the wrong person, but you can fix it after the fact. This is accomplished with the cell of the DataGrid showing the person being made editable with a ComboBox appearing that allows you to select any other person from the database.
- Initial start order of the events in the event tab of TrailMeisterViewer is now reverse sorted by date so the most recent is at the top.
- Some alignment fixes for column headers in the viewer
- Fixed a log of possible null refs and other warnings in the IDE; all clean now.

+++ TrailMeister/TrailMeister/GUI/Main/MainWindowVM.cs
@@ -18,7 +18,7 @@ namespace TrailMeister.GUI.Main
         private ObservableKeyedCollection<int, RecentLapData> _recentLapData = new ObservableKeyedCollection<int, RecentLapData>(null, "ID");
         private RecentLapData? _mostRecentData;
         private ObservableKeyedCollection<string, RecentLapData> _allParticipants = new ObservableKeyedCollection<string, RecentLapData>(null, "EPC");
-        private ObservableKeyedCollection<int, DbPerson> _allPeople = new ObservableKeyedCollection<int, DbPerson>(null, "PersonId");
+        private ObservableKeyedCollection<long, DbPerson> _allPeople = new ObservableKeyedCollection<long, DbPerson>(null, "PersonId");
         protected readonly MainWindowController _controller;
 
         private string _eventName = "Event " + DateOnly.FromDateTime(DateTime.Today).ToString("yyyy/MM/dd");
@@ -171,7 +171,7 @@ namespace TrailMeister.GUI.Main
             }
         }
 
-        public ObservableKeyedCollection<int, DbPerson> AllPeople
+        public ObservableKeyedCollection<long, DbPerson> AllPeople
         {
             get
             {
